ViewModelFactory separated from MovieDetailsViewModel
